@@ -1,20 +1,21 @@
 import chalk from "chalk";
 import { execa } from "execa";
 import ora, { type Ora } from "ora";
+
 import {
   getUserPkgManager,
   type PackageManager,
 } from "~/utils/getUserPkgManager.js";
 import { logger } from "~/utils/logger.js";
 
-type Options = {
+interface Options {
   projectDir: string;
-};
+}
 
 /*eslint-disable @typescript-eslint/no-floating-promises*/
 const runInstallCommand = async (
   pkgManager: PackageManager,
-  projectDir: string,
+  projectDir: string
 ): Promise<Ora | null> => {
   switch (pkgManager) {
     // When using npm, inherit the stderr stream so that the progress bar is shown
@@ -76,7 +77,7 @@ export const installDependencies = async ({ projectDir }: Options) => {
 
   // If the spinner was used to show the progress, use succeed method on it
   // If not, use the succeed on a new spinner
-  (installSpinner || ora()).succeed(
-    chalk.green("Successfully installed dependencies!\n"),
+  (installSpinner ?? ora()).succeed(
+    chalk.green("Successfully installed dependencies!\n")
   );
 };
