@@ -1,14 +1,15 @@
 import chalk from "chalk";
 import ora from "ora";
+
 import {
   type InstallerOptions,
   type PkgInstallerMap,
 } from "~/installers/index.js";
 import { logger } from "~/utils/logger.js";
 
-type InstallPackagesOptions = {
+type InstallPackagesOptions = InstallerOptions & {
   packages: PkgInstallerMap;
-} & InstallerOptions;
+};
 // This runs the installer for all the packages that the user has selected
 export const installPackages = (options: InstallPackagesOptions) => {
   const { packages } = options;
@@ -20,8 +21,8 @@ export const installPackages = (options: InstallPackagesOptions) => {
       pkgOpts.installer(options);
       spinner.succeed(
         chalk.green(
-          `Successfully setup boilerplate for ${chalk.green.bold(name)}`,
-        ),
+          `Successfully setup boilerplate for ${chalk.green.bold(name)}`
+        )
       );
     }
   }
