@@ -1,8 +1,9 @@
+import path from "path";
+import * as p from "@clack/prompts";
 import chalk from "chalk";
 import fs from "fs-extra";
-import inquirer from "inquirer";
 import ora from "ora";
-import path from "path";
+
 import { PKG_ROOT } from "~/consts.js";
 import { type InstallerOptions } from "~/installers/index.js";
 import { logger } from "~/utils/logger.js";
@@ -28,38 +29,29 @@ export const scaffoldProject = async ({
     if (fs.readdirSync(projectDir).length === 0) {
       if (projectName !== ".")
         spinner.info(
-          `${chalk.cyan.bold(
-            projectName,
-          )} exists but is empty, continuing...\n`,
+          `${chalk.cyan.bold(projectName)} exists but is empty, continuing...\n`
         );
     } else {
       spinner.stopAndPersist();
-      const { overwriteDir } = await inquirer.prompt<{
-        overwriteDir: "abort" | "clear" | "overwrite";
-      }>({
-        name: "overwriteDir",
-        type: "list",
+      const overwriteDir = await p.select({
         message: `${chalk.redBright.bold("Warning:")} ${chalk.cyan.bold(
-          projectName,
+          projectName
         )} already exists and isn't empty. How would you like to proceed?`,
-        choices: [
+        options: [
           {
-            name: "Abort installation (recommended)",
+            label: "Abort installation (recommended)",
             value: "abort",
-            short: "Abort",
           },
           {
-            name: "Clear the directory and continue installation",
+            label: "Clear the directory and continue installation",
             value: "clear",
-            short: "Clear",
           },
           {
-            name: "Continue installation and overwrite conflicting files",
+            label: "Continue installation and overwrite conflicting files",
             value: "overwrite",
-            short: "Overwrite",
           },
         ],
-        default: "abort",
+        initialValue: "abort",
       });
       if (overwriteDir === "abort") {
         spinner.fail("Aborting installation...");
@@ -71,13 +63,9 @@ export const scaffoldProject = async ({
           ? "clear the directory"
           : "overwrite conflicting files";
 
-      const { confirmOverwriteDir } = await inquirer.prompt<{
-        confirmOverwriteDir: boolean;
-      }>({
-        name: "confirmOverwriteDir",
-        type: "confirm",
+      const confirmOverwriteDir = await p.confirm({
         message: `Are you sure you want to ${overwriteAction}?`,
-        default: false,
+        initialValue: false,
       });
 
       if (!confirmOverwriteDir) {
@@ -87,7 +75,7 @@ export const scaffoldProject = async ({
 
       if (overwriteDir === "clear") {
         spinner.info(
-          `Emptying ${chalk.cyan.bold(projectName)} and creating t3 app..\n`,
+          `Emptying ${chalk.cyan.bold(projectName)} and creating t3 app..\n`
         );
         fs.emptyDirSync(projectDir);
       }
@@ -99,13 +87,13 @@ export const scaffoldProject = async ({
   fs.copySync(srcDir, projectDir);
   fs.renameSync(
     path.join(projectDir, "_gitignore"),
-    path.join(projectDir, ".gitignore"),
+    path.join(projectDir, ".gitignore")
   );
 
   const scaffoldedName =
     projectName === "." ? "App" : chalk.cyan.bold(projectName);
 
   spinner.succeed(
-    `${scaffoldedName} ${chalk.green("scaffolded successfully!")}\n`,
+    `${scaffoldedName} ${chalk.green("scaffolded successfully!")}\n`
   );
 };
