@@ -1,6 +1,7 @@
-import fs from "fs-extra";
 import path from "path";
+import fs from "fs-extra";
 import { type PackageJson } from "type-fest";
+
 import { PKG_ROOT } from "~/consts.js";
 import { type Installer } from "~/installers/index.js";
 import { addPackageDependency } from "~/utils/addPackageDependency.js";
@@ -22,7 +23,7 @@ export const prismaInstaller: Installer = ({ projectDir, packages }) => {
   const schemaSrc = path.join(
     extrasDir,
     "prisma/schema",
-    packages?.nextAuth.inUse ? "with-auth.prisma" : "base.prisma",
+    packages?.nextAuth.inUse ? "with-auth.prisma" : "base.prisma"
   );
   const schemaDest = path.join(projectDir, "prisma/schema.prisma");
 
