@@ -1,7 +1,8 @@
-import fs from "fs-extra";
 import path from "path";
+import fs from "fs-extra";
 import sortPackageJson from "sort-package-json";
 import { type PackageJson } from "type-fest";
+
 import {
   dependencyVersionMap,
   type AvailableDependencies,
@@ -15,7 +16,7 @@ export const addPackageDependency = (opts: {
   const { dependencies, devMode, projectDir } = opts;
 
   const pkgJson = fs.readJSONSync(
-    path.join(projectDir, "package.json"),
+    path.join(projectDir, "package.json")
   ) as PackageJson;
 
   dependencies.forEach((pkgName) => {
