@@ -1,7 +1,8 @@
-import { getVersion } from "./getT3Version.js";
-import { logger } from "./logger.js";
 import { execSync } from "child_process";
 import https from "https";
+
+import { getVersion } from "./getT3Version.js";
+import { logger } from "./logger.js";
 
 export const renderVersionWarning = (npmVersion: string) => {
   const currentVersion = getVersion();
@@ -14,7 +15,7 @@ export const renderVersionWarning = (npmVersion: string) => {
     logger.warn("  Please report any bugs you encounter.");
   } else if (currentVersion.includes("next")) {
     logger.warn(
-      "  You are running create-t3-app with the @next tag which is no longer maintained.",
+      "  You are running create-t3-app with the @next tag which is no longer maintained."
     );
     logger.warn("  Please run the CLI with @latest instead.");
   } else if (currentVersion !== npmVersion) {
@@ -23,7 +24,7 @@ export const renderVersionWarning = (npmVersion: string) => {
       "  Your version:",
       currentVersion + ".",
       "Latest version in the npm registry:",
-      npmVersion,
+      npmVersion
     );
     logger.warn("  Please run the CLI with @latest to get the latest updates.");
   }
@@ -37,9 +38,9 @@ export const renderVersionWarning = (npmVersion: string) => {
  * directory of this source tree.
  * https://github.com/facebook/create-react-app/blob/main/packages/create-react-app/LICENSE
  */
-type DistTagsBody = {
+interface DistTagsBody {
   latest: string;
-};
+}
 
 function checkForLatestVersion(): Promise<string> {
   return new Promise((resolve, reject) => {
@@ -56,7 +57,7 @@ function checkForLatestVersion(): Promise<string> {
           } else {
             reject();
           }
-        },
+        }
       )
       .on("error", () => {
         // logger.error("Unable to check for latest version.");
