@@ -1,8 +1,11 @@
+import { removeTrailingSlash } from "./removeTrailingSlash.js";
+
 const validationRegExp =
   /^(?:@[a-z0-9-*~][a-z0-9-*._~]*\/)?[a-z0-9-~][a-z0-9-._~]*$/;
 
 //Validate a string against allowed package.json names
-export const validateAppName = (input: string) => {
+export const validateAppName = (rawInput: string) => {
+  const input = removeTrailingSlash(rawInput);
   const paths = input.split("/");
 
   // If the first part is a @, it's a scoped package
@@ -14,7 +17,7 @@ export const validateAppName = (input: string) => {
   }
 
   if (input === "." || validationRegExp.test(appName ?? "")) {
-    return true;
+    return;
   } else {
     return "App name must consist of only lowercase alphanumeric characters, '-', and '_'";
   }
