@@ -1,9 +1,14 @@
 import * as dotenv from "dotenv";
 import { type Config } from "drizzle-kit";
 
+import { env } from "~/env.mjs";
+
 dotenv.config();
 
 export default {
-  schema: "./src/server/db/schema.ts",
-  connectionString: process.env.DATABASE_URL,
+  schema: "./schema",
+  driver: "mysql2",
+  dbCredentials: {
+    connectionString: env.DATABASE_URL,
+  },
 } satisfies Config;
