@@ -1,15 +1,15 @@
-import { and, eq } from "drizzle-orm";
+import { DrizzleAdapter } from "@auth/drizzle-adapter";
 import { type GetServerSidePropsContext } from "next";
 import {
   getServerSession,
-  type NextAuthOptions,
   type DefaultSession,
+  type NextAuthOptions,
 } from "next-auth";
-import { type Adapter } from "next-auth/adapters";
 import DiscordProvider from "next-auth/providers/discord";
+
 import { env } from "~/env.mjs";
 import { db } from "~/server/db";
-import * as schema from "~/server/db/schema";
+import { mysqlTable } from "~/server/db/schema";
 
 /**
  * Module augmentation for `next-auth` types. Allows us to add custom properties to the `session`
@@ -47,7 +47,7 @@ export const authOptions: NextAuthOptions = {
       },
     }),
   },
-  adapter: DrizzleAdapter(),
+  adapter: DrizzleAdapter(db, mysqlTable),
   providers: [
     DiscordProvider({
       clientId: env.DISCORD_CLIENT_ID,
@@ -76,165 +76,3 @@ export const getServerAuthSession = (ctx: {
 }) => {
   return getServerSession(ctx.req, ctx.res, authOptions);
 };
-
-/**
- * Adapter for Drizzle ORM. This is not yet available in NextAuth directly, so we inhouse our own.
- * When the official one is out, we will switch to that.
- *
- * @see
- * https://github.com/nextauthjs/next-auth/pull/7165/files#diff-142e7d6584eed63a73316fbc041fb93a0564a1cbb0da71200b92628ca66024b5
- */
-
-export function DrizzleAdapter(): Adapter {
-  const { users, sessions, accounts, verificationTokens } = schema;
-  return {
-    createUser: async (data) => {
-      const id = crypto.randomUUID();
-
-      await db.insert(users).values({ ...data, id });
-
-      const user = await db
-        .select()
-        .from(users)
-        .where(eq(users.id, id))
-        .then((res) => res[0]);
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      return user!;
-    },
-    getUser: async (data) => {
-      const user = await db.select().from(users).where(eq(users.id, data));
-      return user[0] ?? null;
-    },
-    getUserByEmail: async (data) => {
-      const user = await db.select().from(users).where(eq(users.email, data));
-      return user[0] ?? null;
-    },
-    createSession: async (data) => {
-      await db.insert(sessions).values(data);
-
-      const session = await db
-        .select()
-        .from(sessions)
-        .where(eq(sessions.sessionToken, data.sessionToken));
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      return session[0]!;
-    },
-    getSessionAndUser: async (data) => {
-      const sessionAndUser = await db
-        .select({
-          session: sessions,
-          user: users,
-        })
-        .from(sessions)
-        .where(eq(sessions.sessionToken, data))
-        .innerJoin(users, eq(users.id, sessions.userId));
-
-      return sessionAndUser[0] ?? null;
-    },
-    updateUser: async (data) => {
-      if (!data.id) {
-        throw new Error("No user id.");
-      }
-
-      await db.update(users).set(data).where(eq(users.id, data.id));
-
-      const user = await db.select().from(users).where(eq(users.id, data.id));
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      return user[0]!;
-    },
-    updateSession: async (data) => {
-      await db
-        .update(sessions)
-        .set(data)
-        .where(eq(sessions.sessionToken, data.sessionToken));
-
-      return db
-        .select()
-        .from(sessions)
-        .where(eq(sessions.sessionToken, data.sessionToken))
-        .then((res) => res[0]);
-    },
-    linkAccount: async (rawAccount) => {
-      await db
-        .insert(accounts)
-        .values(rawAccount)
-        .then((res) => res.rows[0]);
-    },
-    getUserByAccount: async (account) => {
-      const dbAccount = await db
-        .select()
-        .from(accounts)
-        .where(
-          and(
-            eq(accounts.providerAccountId, account.providerAccountId),
-            eq(accounts.provider, account.provider),
-          ),
-        )
-        .leftJoin(users, eq(accounts.userId, users.id))
-        .then((res) => res[0]);
-
-      return dbAccount?.users ?? null;
-    },
-    deleteSession: async (sessionToken) => {
-      await db.delete(sessions).where(eq(sessions.sessionToken, sessionToken));
-    },
-    createVerificationToken: async (token) => {
-      await db.insert(verificationTokens).values(token);
-
-      return db
-        .select()
-        .from(verificationTokens)
-        .where(eq(verificationTokens.identifier, token.identifier))
-        .then((res) => res[0]);
-    },
-    useVerificationToken: async (token) => {
-      try {
-        const deletedToken =
-          (await db
-            .select()
-            .from(verificationTokens)
-            .where(
-              and(
-                eq(verificationTokens.identifier, token.identifier),
-                eq(verificationTokens.token, token.token),
-              ),
-            )
-            .then((res) => res[0])) ?? null;
-
-        await db
-          .delete(verificationTokens)
-          .where(
-            and(
-              eq(verificationTokens.identifier, token.identifier),
-              eq(verificationTokens.token, token.token),
-            ),
-          );
-
-        return deletedToken;
-      } catch (err) {
-        throw new Error("No verification token found.");
-      }
-    },
-    deleteUser: async (id) => {
-      await Promise.all([
-        db.delete(users).where(eq(users.id, id)),
-        db.delete(sessions).where(eq(sessions.userId, id)),
-        db.delete(accounts).where(eq(accounts.userId, id)),
-      ]);
-
-      return null;
-    },
-    unlinkAccount: async (account) => {
-      await db
-        .delete(accounts)
-        .where(
-          and(
-            eq(accounts.providerAccountId, account.providerAccountId),
-            eq(accounts.provider, account.provider),
-          ),
-        );
-
-      return undefined;
-    },
-  };
-}
