@@ -1,6 +1,7 @@
 // Example model schema from the Drizzle docs
 // https://orm.drizzle.team/docs/sql-schema-declaration
 
+import { sql } from "drizzle-orm";
 import {
   mysqlTableCreator,
   serial,
@@ -15,17 +16,17 @@ import {
  *
  * @see https://orm.drizzle.team/docs/goodies#multi-project-schema
  */
-const mysqlTable = mysqlTableCreator((name) => `project1_${name}`);
+export const mysqlTable = mysqlTableCreator((name) => `project1_${name}`);
 
 export const example = mysqlTable(
   "example",
   {
     id: serial("id").primaryKey(),
     name: varchar("name", { length: 256 }),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
+    createdAt: timestamp("created_at").default(sql`CURRENT_TIMESTAMP`).notNull(),
     updatedAt: timestamp("updatedAt").onUpdateNow(),
   },
   (example) => ({
     nameIndex: uniqueIndex("name_idx").on(example.name),
-  }),
+  })
 );
