@@ -1,11 +1,12 @@
-import * as schema from "./schema";
-import { connect } from "@planetscale/database";
+import { Client } from "@planetscale/database";
 import { drizzle } from "drizzle-orm/planetscale-serverless";
+
 import { env } from "~/env.mjs";
+import * as schema from "./schema";
 
-// create the connection
-const connection = connect({
-  url: env.DATABASE_URL,
-});
-
-export const db = drizzle(connection, { schema });
+export const db = drizzle(
+  new Client({
+    url: env.DATABASE_URL,
+  }).connection(),
+  { schema }
+);
