@@ -1,6 +1,7 @@
 import { Slot } from "@radix-ui/react-slot";
-import { type VariantProps, cva } from "class-variance-authority";
+import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
+
 import { cn } from "~/lib/utils";
 
 const buttonVariants = cva(
