@@ -1,8 +1,9 @@
 "use client";
 
 import * as LabelPrimitive from "@radix-ui/react-label";
-import { type VariantProps, cva } from "class-variance-authority";
+import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
+
 import { cn } from "~/lib/utils";
 
 const labelVariants = cva(
