@@ -1,12 +1,13 @@
-import { type KnownLanguageCode, KNOWN_LANGUAGES } from "../../config";
-import { getIsRtlFromLangCode } from "../../languages";
 import { Listbox, Transition } from "@headlessui/react";
 import clsx from "clsx";
 import { Fragment } from "react";
 
-type LanguageSelectProps = {
+import { KNOWN_LANGUAGES, type KnownLanguageCode } from "../../config";
+import { getIsRtlFromLangCode } from "../../languages";
+
+interface LanguageSelectProps {
   language: KnownLanguageCode;
-};
+}
 
 export default function LanguageSelect({ language }: LanguageSelectProps) {
   const handleSelect = (code: string) => {
