@@ -1,8 +1,11 @@
 /** @jsxImportSource react */
 
+import { useCallback, useRef, useState } from "react";
+
 import { ALGOLIA } from "../../config";
-import { useState, useCallback, useRef } from "react";
+
 import "../../styles/algolia/style.css";
+
 import * as docSearchReact from "@docsearch/react";
 import clsx from "clsx";
 import { createPortal } from "react-dom";
@@ -10,9 +13,11 @@ import { createPortal } from "react-dom";
 /** FIXME: This is still kinda nasty, but DocSearch is not ESM ready. */
 const DocSearchModal =
   docSearchReact.DocSearchModal ||
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access
   (docSearchReact as any).default.DocSearchModal;
 const useDocSearchKeyboardEvents =
   docSearchReact.useDocSearchKeyboardEvents ||
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access
   (docSearchReact as any).default.useDocSearchKeyboardEvents;
 
 export default function Search({ isLanding }: { isLanding: boolean }) {
