@@ -1,11 +1,12 @@
+import { Resvg } from "@resvg/resvg-js";
+import { type APIRoute } from "astro";
+import satori from "satori";
+
 import OpenGraph from "../components/openGraph";
 import { SITE } from "../config";
 import { getIsRtlFromLangCode, getLanguageFromURL } from "../languages";
 import { getFont } from "../utils/ogFont";
 import { SITE_URL } from "../utils/siteUrl";
-import { Resvg } from "@resvg/resvg-js";
-import { type APIRoute } from "astro";
-import satori from "satori";
 
 const removeEndingSlash = (str: string) => str.replace(/\/$/, "");
 
